Add API functions
fetchData(): Is the FETCH request returns a json
response from the api using axios. Needs to be parsed
into a HTML row for the table

submitData(e): is the POST request that takes an
event 'e' parameter from the form submission and
logs the POST response if successful.

deleteData(): takes an event 'e' parameter because
it will be attached to the Delete <button> on each
row. Needs the ID for the data fetched and parsed
into the row to include in the API url